- Removed Material id system to add 1.13 support
notes:
this means that old chest and kit configs will need to be edited to use
the material name system(ex: 267 to iron_sword).
the config.yml block break/place whitelist will also need to be updated
make sure to use the correct name on 1.12 (so YELLOW_FLOWER instead of
DANDELION)on 1.13 both with work(Waring: it won't convert wool,14 into
RED_WOOL so i recommend you use the 1.13 name if you're on 1.13).